add static content

+++ src/main/java/com/ozhegov/cloudstorage/controllers/FilesController.java
@@ -5,11 +5,17 @@ import com.ozhegov.cloudstorage.exception.FileIsAlreadyExistsException;
 import com.ozhegov.cloudstorage.exception.NoSuchFileException;
 import com.ozhegov.cloudstorage.dto.Blob;
 import com.ozhegov.cloudstorage.dto.ErrorMessage;
+import com.ozhegov.cloudstorage.model.StorageUser;
+import com.ozhegov.cloudstorage.repository.UserRepository;
 import com.ozhegov.cloudstorage.service.FileService;
 import io.minio.errors.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.*;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -17,14 +23,17 @@ import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/v1/storage")
+@RequestMapping("/api")
 public class FilesController {
     @Autowired
     private FileService service;
-    @PostMapping("/directory?path={path}")
-    public ResponseEntity<String> createDirectory(@PathVariable String path){
+    @Autowired
+    private UserRepository userRepository;
+    @PostMapping("/directory")
+    public ResponseEntity<String> createDirectory(@RequestParam String path){
         try {
             Blob blob = service.createDirectory(path);
             String json = (new Gson()).toJson(blob);
@@ -39,9 +48,13 @@ public class FilesController {
             throw new RuntimeException(e);
         }
     }
-    @GetMapping("/directory?path={path}")
-    public ResponseEntity<String> getFilesInDir(@PathVariable String path){
+    @GetMapping("/directory")
+    public ResponseEntity<String> getFilesInDir(@RequestParam String path){
         try {
+            Long userId = getCurrentUserId();
+            if(userId == null)
+                return ResponseEntity.status(401).body("Сессия пользователя истекла");
+            path = "user-" + userId + "-files/" + path;
             List<Blob> files = service.findAllInDir(path);
             String json = (new Gson()).toJson(files);
             return ResponseEntity.status(200).body(json);
@@ -52,8 +65,12 @@ public class FilesController {
             throw new RuntimeException(e);
         }
     }
-    @GetMapping("/resource?path={path}")
-    public ResponseEntity<String> getFileByPath(@PathVariable String path){
+    @GetMapping("/resource")
+    public ResponseEntity<String> getFileByPath(@RequestParam String path){
+        Long userId = getCurrentUserId();
+        if(userId == null)
+            return ResponseEntity.status(401).body("Сессия пользователя истекла");
+        path = "user-" + userId + "-files/" + path;
         try {
             Blob blob = service.getFile(path);
             String json = (new Gson()).toJson(blob);
@@ -70,8 +87,14 @@ public class FilesController {
             throw new RuntimeException(e);
         }
     }
-    @GetMapping("/resource/download?path={path}")
-    public ResponseEntity<Resource> downloadFile(@PathVariable String path){
+    @GetMapping("/resource/download")
+    public ResponseEntity<Resource> downloadFile(@RequestParam String path){
+        Long userId = getCurrentUserId();
+        if(userId == null) {
+            ByteArrayResource empty = new ByteArrayResource(new byte[0]);
+            return ResponseEntity.status(401).body(empty);
+        }
+        path = "user-" + userId + "-files/" + path;
         try{
             Resource resource = service.downloadFile(path);
 
@@ -85,8 +108,12 @@ public class FilesController {
             throw new RuntimeException();
         }
     }
-    @PostMapping("/resource?path={path}")
-    public ResponseEntity<String> uploadFile(@PathVariable String path, @RequestParam MultipartFile file){
+    @PostMapping("/resource")
+    public ResponseEntity<String> uploadFile(@RequestParam String path, @RequestParam MultipartFile file){
+        Long userId = getCurrentUserId();
+        if(userId == null)
+            return ResponseEntity.status(401).body("Сессия пользователя истекла");
+        path = "user-" + userId + "-files/" + path;
         try{
             Blob blob = service.uploadFile(file, path);
             String json = (new Gson()).toJson(blob);
@@ -100,8 +127,13 @@ public class FilesController {
             return ResponseEntity.status(409).body(json);
         }
     }
-    @GetMapping("/resource/move?from={from}&to={to}")
-    public ResponseEntity<String> replaceResource(@PathVariable String from, @PathVariable String to){
+    @GetMapping("/resource/move")
+    public ResponseEntity<String> replaceResource(@RequestParam String from, @RequestParam String to){
+        Long userId = getCurrentUserId();
+        if(userId == null)
+            return ResponseEntity.status(401).body("Сессия пользователя истекла");
+        to = "user-" + userId + "-files/" + to;
+        from = "user-" + userId + "-files/" + from;
         try{
             Blob blob = service.replaceResource(from, to);
             String json = (new Gson()).toJson(blob);
@@ -115,17 +147,28 @@ public class FilesController {
             throw new RuntimeException(e);
         }
     }
-    @DeleteMapping("/resource?path={path}")
-    public HttpStatus deleteFileByPath(@PathVariable String path) {
+    @DeleteMapping("/resource")
+    public ResponseEntity<String> deleteFileByPath(@RequestParam String path) {
+        Long userId = getCurrentUserId();
+        if(userId == null)
+            return ResponseEntity.status(401).body("Сессия пользователя истекла");
+        path = "user-" + userId + "-files/" + path;
         try {
             service.deleteFile(path);
-            return HttpStatus.valueOf(204);
+            return ResponseEntity.status(204).body("Ресурс успешно удален");
         }catch(NoSuchFileException e){
-            return HttpStatus.valueOf(404);
+            return ResponseEntity.status(404).body("Указанный ресурс не найден");
         }catch (ServerException | InsufficientDataException | IOException |
                 NoSuchAlgorithmException | InvalidKeyException | InvalidResponseException | XmlParserException |
                 InternalException | ErrorResponseException e) {
-            return HttpStatus.INTERNAL_SERVER_ERROR;
+            return ResponseEntity.status(500).body("Ошибка сервера");
         }
+    }
+
+    private Long getCurrentUserId(){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = auth.getName();
+        Optional<StorageUser> user = userRepository.findByName(username);
+        return user.map(StorageUser::getId).orElse(null);
     }
 }
